requirements.txt
# OCR & Image Processing
pytesseract==0.3.10
pdf2image==1.16.3
opencv-python==4.8.0.76
numpy==1.24.3
Pillow==10.0.0

# Data Handling
pandas
spacy